+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["gonk", "gonk-core", "gonk-player", "gonk-database"]
+members = ["gonk", "gonk-core", "gonk-player"]
 
 [profile.release]
 strip = true
