+++ gonk-core/src/lazy.rs
