+++ gonk-core/src/vdb.rs
