--- gonk-database/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "gonk-database"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-rayon = "1.6.0"
-walkdir = "2.3.2"
